refactor(peripheral-demo): rename TS files to kebab-case and simplify CSS comments
- Rename screenEvents, operatorDashboard, publicClock, safetyCarousel
- Update all imports to match new file names
- Replace decorated CSS section comments with simple single-line comments
