+++ edge-apps/peripheral-integration-demo/src/core/screen-events.ts
