+++ edge-apps/peripheral-integration-demo/src/features/operator-dashboard.ts
