+++ edge-apps/peripheral-integration-demo/src/features/public-clock.ts
