+++ edge-apps/peripheral-integration-demo/src/features/safety-carousel.ts
